requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.0
einops>=0.4.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.8.0
mlflow>=1.24.0
wandb>=0.12.0
pyyaml>=6.0
albumentations>=1.2.0
# Phase 3 Advanced Features
optuna>=3.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
umap-learn>=0.5.0
